% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools.R
\name{arrayStrToVec}
\alias{arrayStrToVec}
\title{Write a PostgreSQL array as a string from a vector.}
\usage{
arrayStrToVec(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A string.
}
\description{
Write a PostgreSQL array as a string from a vector.
}
\examples{
arrayStrToVec(vecToArrayStr(c("a", "b")))
}
