% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools.R
\name{connect}
\alias{connect}
\title{Connect to a local database with local credentials using DBI/odbc.}
\usage{
connect(db)
}
\arguments{
\item{db}{A string, a database you can connect to locally.}
}
\value{
A database connection.
}
\description{
Connect to a local database with local credentials using DBI/odbc.
}
\examples{
connect(NULL)
}
