% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_pg.R
\name{connect_pg}
\alias{connect_pg}
\title{Connect to Postgres Database}
\usage{
connect_pg(getenv = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{getenv}{get credentials from the local .REnviron file}

\item{verbose}{print database connection information from \code{\link{dbGetInfo}()}}

\item{...}{other arguments to pass through \code{\link{dbConnect}()}}
}
\value{
connection to Postgres database
}
\description{
This is a convenience wrapper to \code{DBI::dbConnect()} where if the database credentials
are saved in the .REnviron file, they will be automatically read.
}
\details{
All other arguments pass through \code{dbConnect()} except for \code{drv} which is already specified
to \code{RPostgres::Postgres()}
}
\examples{
\dontrun{
connect_pg(getenv = FALSE,
  drv = DBI::dbDriver("Postgres"),
  port = 5432,
  dbname = "mydb",
  user = "myusername",
  password = "mypw"
)
}
}
