% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools.R
\name{createDATABASE}
\alias{createDATABASE}
\title{Generate a PostgreSQL CREATE DATABASE statement, optionally execute the statement if con is not NULL.}
\usage{
createDATABASE(
  name,
  owner = NULL,
  template = NULL,
  encoding = NULL,
  locale = NULL,
  lc_collate = NULL,
  lc_ctype = NULL,
  tablespace = NULL,
  allow_connections = NULL,
  connection_limit = NULL,
  is_template = NULL,
  con = NULL
)
}
\arguments{
\item{name}{A string, the "name" parameter for PostgreSQL CREATE DATABASE.}

\item{owner}{A string, the "user_name" parameter for PostgreSQL CREATE DATABASE.}

\item{template}{A string, the "template" parameter for PostgreSQL CREATE DATABASE.}

\item{encoding}{A string, the "encoding" parameter for PostgreSQL CREATE DATABASE.}

\item{locale}{A string, the "locale" parameter for PostgreSQL CREATE DATABASE}

\item{lc_collate}{A string, the "lc_collate" parameter for PostgreSQL CREATE DATABASE.}

\item{lc_ctype}{A string, the "lc_ctype" parameter for PostgreSQL CREATE DATABASE.}

\item{tablespace}{A string, the "tablespace_name" parameter for PostgreSQL CREATE DATABASE.}

\item{allow_connections}{A string, the "allowconn" parameter for PostgreSQL CREATE DATABASE.}

\item{connection_limit}{A string, the "connlimit" parameter for PostgreSQL CREATE DATABASE.}

\item{is_template}{A string, the "istemplate" parameter for PostgreSQL CREATE DATABASE.}

\item{con}{A database connection that can be passed to DBI::dbSendQuery/DBI::dbGetQuery.}
}
\value{
A string, PostgreSQL CREATE DATABASE statement; or the results retrieved by DBI::dbSendQuery after executing the statement.
}
\description{
Generate a PostgreSQL CREATE DATABASE statement, optionally execute the statement if con is not NULL.
}
\examples{
createDATABASE("dbTest01")
}
