% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools2.R
\name{createEXTENSION}
\alias{createEXTENSION}
\title{Generate a postgreSQL CREATE EXTENSION statement, optionally execute the statement if con is not NULL.}
\usage{
createEXTENSION(
  name,
  if_not_exists = FALSE,
  schema = NULL,
  version = NULL,
  cascade = FALSE,
  con = NULL
)
}
\arguments{
\item{name}{A string, the "extension_name" parameter for postgreSQL CREATE EXTENSION.}

\item{if_not_exists}{TRUE/FALSE, if TRUE, adds "IF NOT EXISTS" to postgreSQL CREATE EXTENSION statement.}

\item{schema}{A string, the "schema_name" parameter for postgreSQL CREATE EXTENSION.}

\item{version}{A string, the "version" parameter for postgreSQL CREATE EXTENSION.}

\item{cascade}{TRUE/FALSE, if TRUE, adds "CASCADE" to postgreSQL CREATE EXTENSION statement.}

\item{con}{A database connection that can be passed to DBI::dbSendQuery/DBI::dbGetQuery.}
}
\value{
A string, postgreSQL CREATE EXTENSION statement; or the results retrieved by DBI::dbSendQuery after executing the statement.
}
\description{
Generate a postgreSQL CREATE EXTENSION statement, optionally execute the statement if con is not NULL.
}
\examples{
createEXTENSION("pgcrypto")
}
