% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools2.R
\name{doubleQuoteText}
\alias{doubleQuoteText}
\title{Add double quotes to strings.}
\usage{
doubleQuoteText(x, char_only = TRUE)
}
\arguments{
\item{x}{A string.}

\item{char_only}{TRUE/FALSE, if TRUE, adds quotes only if is.character(x) is TRUE.}
}
\value{
A string, with double quotes added.
}
\description{
Add double quotes to strings.
}
\examples{
doubleQuoteText("Sample quotes.")
}
