% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nchar.R
\name{get_nchar}
\alias{get_nchar}
\title{Calculate element lengths from a \code{data.frame}}
\usage{
get_nchar(input, export = FALSE, path = NULL)
}
\arguments{
\item{input}{a \code{data.frame} or \code{list} of data frames}

\item{export}{a \code{logical} option export the result as an binary file}

\item{path}{a \code{file path} option to specify the write location of the binary file}
}
\value{
returns a \code{data.frame} or \code{list} of data frames with rows being the input columns and
respective summary statistics and classes for each column
}
\description{
This function calculates the length of every element within a data frame for later use
to map Postgres field types when writing to the database.
}
\details{
The output for the specified summary statistic is based on the \code{nchar()} function.
NOTE: that \code{\link{nchar}()} counts the literal length of an element or the element coerced to a character.
See \code{\link{nchar}()}.
}
\examples{
data(iris)
get_nchar(iris)

}
