% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools.R
\name{insert_batch_chunker}
\alias{insert_batch_chunker}
\title{Helper function for INSERT}
\usage{
insert_batch_chunker(x, n_batches, batch_size)
}
\arguments{
\item{x}{A vector of data to insert.}

\item{n_batches}{Integer, the number of batches needed to insert the data.}

\item{batch_size}{Integer, the size of each batch.}
}
\value{
A list.
}
\description{
Helper function for INSERT
}
\examples{
insert_batch_chunker(c(1, 2, 3), 1, 100)
}
