% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools.R
\name{pg_alterColumnType}
\alias{pg_alterColumnType}
\title{Helper command to alter a column's data type via ALTER TABLE.}
\usage{
pg_alterColumnType(column_name, data_type, using = NULL)
}
\arguments{
\item{column_name}{A string, the name of the column to add.}

\item{data_type}{A string, the data type of the column to add.}

\item{using}{A string, a command to cast the column into the appropriate type.}
}
\value{
A string, PostgreSQL helper statement to alter a column type using ALTER TABLE.
}
\description{
Helper command to alter a column's data type via ALTER TABLE.
}
\examples{
pg_alterColumnType(
column_name = "newCol",
data_type = "text"
)
}
