% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools.R
\name{pg_dropColumn}
\alias{pg_dropColumn}
\title{Helper command to drop a column via ALTER TABLE.}
\usage{
pg_dropColumn(
  column_name,
  cascade = FALSE,
  restrict = FALSE,
  if_exists = FALSE
)
}
\arguments{
\item{column_name}{A string, the name of the column to drop.}

\item{cascade}{Boolean, if TRUE, adds CASCADE to the DROP COLUMN statement.}

\item{restrict}{Boolean, if TRUE, adds RESTRICT to the DROP COLUMN statement.}

\item{if_exists}{Boolean, if TRUE, adds IF EXISTS to the DROP COLUMN statement.}
}
\value{
A string, PostgreSQL helper statement to drop a column using ALTER TABLE.
}
\description{
Helper command to drop a column via ALTER TABLE.
}
\examples{
pg_dropColumn(
column_name = "newCol"
)
}
