% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools.R
\name{pg_renameColumn}
\alias{pg_renameColumn}
\title{Helper command to rename a column via ALTER TABLE.}
\usage{
pg_renameColumn(column_name, new_column_name)
}
\arguments{
\item{column_name}{A string, the name of the column to change.}

\item{new_column_name}{A string, the new name for the column.}
}
\value{
A string, PostgreSQL helper statement to rename a column using ALTER TABLE.
}
\description{
Helper command to rename a column via ALTER TABLE.
}
\examples{
pg_renameColumn(
column_name = "newCol",
new_column_name = "col1"
)
}
