% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools.R
\name{pg_renameTable}
\alias{pg_renameTable}
\title{Helper command to rename a table via ALTER TABLE.}
\usage{
pg_renameTable(new_table_name)
}
\arguments{
\item{new_table_name}{A string, the new name for the table.}
}
\value{
A string, PostgreSQL helper statement to rename a table using ALTER TABLE.
}
\description{
Helper command to rename a table via ALTER TABLE.
}
\examples{
pg_renameTable(
new_table_name = "table1"
)
}
