% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools2.R
\name{quoteText}
\alias{quoteText}
\title{Add single quotes to strings, useful for converting R strings into SQL formatted strings.}
\usage{
quoteText(x, char_only = TRUE)
}
\arguments{
\item{x}{A string.}

\item{char_only}{TRUE/FALSE, if TRUE, adds quotes only if is.character(x) is TRUE.}
}
\value{
A string, with single quotes added to match postgreSQL string formatting.
}
\description{
Add single quotes to strings, useful for converting R strings into SQL formatted strings.
}
\examples{
quoteText("Sample quotes.")
}
