% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools2.R
\name{sampleStr}
\alias{sampleStr}
\title{Generates (pseudo)random strings of the specified char length.
Used in INSERT and UPDATE to avoid overwriting existing columns of data.tables.
Not as fast as stringi:stri_rand_strings.}
\usage{
sampleStr(char)
}
\arguments{
\item{char}{A integer, the number of chars to include in the output string.}
}
\value{
A string.
}
\description{
Generates (pseudo)random strings of the specified char length.
Used in INSERT and UPDATE to avoid overwriting existing columns of data.tables.
Not as fast as stringi:stri_rand_strings.
}
\examples{
sampleStr(10)
}
