% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools.R
\name{sql_80_char_comment}
\alias{sql_80_char_comment}
\title{Add a 80 char SQL comment, intended to be used for visual breaks in documents.}
\usage{
sql_80_char_comment()
}
\value{
A string, 80 chars of "-".
}
\description{
Add a 80 char SQL comment, intended to be used for visual breaks in documents.
}
\examples{
sql_80_char_comment()
}
