% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools2.R
\name{sql_comment}
\alias{sql_comment}
\title{Add a single line SQL comment.}
\usage{
sql_comment(x)
}
\arguments{
\item{x}{A string.}
}
\value{
A string prefixed with "--".
}
\description{
Add a single line SQL comment.
}
\examples{
sql_comment("Sample single line comment.")
}
