% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools.R
\name{sqlTypeWalk}
\alias{sqlTypeWalk}
\title{Get the PostgreSQL data type for a given R data type.}
\usage{
sqlTypeWalk(x)
}
\arguments{
\item{x}{A string, a R data type.}
}
\value{
A string, the PostgreSQL data type for x.
}
\description{
Get the PostgreSQL data type for a given R data type.
}
\examples{
sqlTypeWalk(100.1209)
}
