% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pgtable.R
\name{test_pgtable}
\alias{test_pgtable}
\title{Test Recent Table Write to Postgres}
\usage{
test_pgtable(x, y, order_var = NULL)
}
\arguments{
\item{x}{the original table}

\item{y}{the imported table residing in Postgres}

\item{order_var}{a \code{character} string specifying variable to order both tables}
}
\value{
console output of tests, see \code{Description}
}
\description{
Test Recent Table Write to Postgres
}
\examples{
\dontrun{
1 + 1
}
}
