% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools.R
\name{vecToArrayStr2}
\alias{vecToArrayStr2}
\title{Write a PostgreSQL array as a string using {} format from a vector.}
\usage{
vecToArrayStr2(x, double_quote = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{double_quote}{TRUE/FALSE, if TRUE, the elements of x will be double quoted.}
}
\value{
A string.
}
\description{
Write a PostgreSQL array as a string using {} format from a vector.
}
\examples{
vecToArrayStr2(c("a", "b"))
}
