% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools.R
\name{alterSCHEMA}
\alias{alterSCHEMA}
\title{Generate a PostgreSQL ALTER SCHEMA statement, optionally execute the statement if con is not NULL.}
\usage{
alterSCHEMA(name, rename_to = NULL, owner_to = NULL, con = NULL)
}
\arguments{
\item{name}{A string, the "name" parameter for PostgreSQL ALTER SCHEMA.}

\item{rename_to}{A string, the "new_name" parameter for PostgreSQL ALTER SCHEMA.}

\item{owner_to}{A string, the "new_owner" parameter for PostgreSQL ALTER SCHEMA.}

\item{con}{A database connection that can be passed to DBI::dbSendQuery/DBI::dbGetQuery.}
}
\value{
A string, PostgreSQL CREATE SCHEMA statement; or the results retrieved by DBI::dbSendQuery after executing the statement.
}
\description{
Generate a PostgreSQL ALTER SCHEMA statement, optionally execute the statement if con is not NULL.
}
\examples{
alterSCHEMA("dev", rename_to = "prod")
}
