% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools2.R
\name{dropEXTENSION}
\alias{dropEXTENSION}
\title{Generate a postgreSQL DROP EXTENSION statement, optionally execute the statement if con is not NULL.}
\usage{
dropEXTENSION(
  name,
  if_exists = FALSE,
  cascade = FALSE,
  restrict = FALSE,
  con = NULL
)
}
\arguments{
\item{name}{A string, the "name" parameter for postgreSQL DROP EXTENSION.}

\item{if_exists}{TRUE/FALSE, if TRUE, adds "IF EXISTS" to postgreSQL DROP EXTENSION statement.}

\item{cascade}{TRUE/FALSE, if TRUE, adds "CASCADE" to postgreSQL DROP EXTENSION statement.}

\item{restrict}{TRUE/FALSE, if TRUE, adds "RESTRICT" to postgreSQL DROP EXTENSION statement.}

\item{con}{A database connection that can be passed to DBI::dbSendQuery/DBI::dbGetQuery.}
}
\value{
A string, postgreSQL DROP EXTENSION statement; or the results retrieved by DBI::dbSendQuery after executing the statement.
}
\description{
Generate a postgreSQL DROP EXTENSION statement, optionally execute the statement if con is not NULL.
}
\examples{
dropEXTENSION("pgcrypto")
}
