% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgTools.R
\name{vecToArrayStr}
\alias{vecToArrayStr}
\title{Write a PostgreSQL array as a string using ARRAY[] format from a vector.}
\usage{
vecToArrayStr(x, quote = TRUE)
}
\arguments{
\item{x}{A vector.}

\item{quote}{TRUE/FALSE, if TRUE, the elements of x will be quoted.}
}
\value{
A string.
}
\description{
Write a PostgreSQL array as a string using ARRAY[] format from a vector.
}
\examples{
vecToArrayStr(c("a", "b"))
}
